refactor(models): improve typing and tests
