+++ packages/models/src/index.ts
@@ -1,4 +1,8 @@
-export { CategoryConfig, categoryConfigSchema } from './lib/category-config';
+export {
+  CategoryConfigRef,
+  CategoryConfig,
+  categoryConfigSchema,
+} from './lib/category-config';
 export {
   CoreConfig,
   coreConfigSchema,
@@ -12,20 +16,20 @@ export {
   formatSchema,
   persistConfigSchema,
 } from './lib/persist-config';
+export { PluginConfig, pluginConfigSchema } from './lib/plugin-config';
+export { RunnerConfig } from './lib/plugin-config-runner';
 export {
-  Audit,
-  AuditGroup,
-  AuditOutput,
-  AuditOutputs,
-  Issue,
-  PluginConfig,
-  auditGroupSchema,
-  auditOutputsSchema,
   auditSchema,
-  issueSchema,
-  pluginConfigSchema,
-  RunnerConfig,
-} from './lib/plugin-config';
+  Audit,
+  pluginAuditsSchema,
+} from './lib/plugin-config-audits';
+export {
+  AuditGroupRef,
+  AuditGroup,
+  auditGroupSchema,
+} from './lib/plugin-config-groups';
+export { AuditOutput, auditOutputsSchema } from './lib/plugin-process-output';
+export { Issue } from './lib/plugin-process-output-audit-issue';
 export {
   AuditReport,
   PluginReport,
@@ -34,3 +38,4 @@ export {
   reportSchema,
 } from './lib/report';
 export { UploadConfig, uploadConfigSchema } from './lib/upload-config';
+export { materialIconSchema } from './lib/implementation/schemas';
