+++ packages/models/src/lib/category-config.ts
@@ -7,28 +7,25 @@ import {
 } from './implementation/schemas';
 import { errorItems, hasDuplicateStrings } from './implementation/utils';
 
-type _RefsList = {
-  type?: string;
-  slug?: string;
-  plugin?: string;
-}[];
+export const categoryConfigRefSchema = weightedRefSchema(
+  'Weighted references to audits and/or groups for the category',
+  'Slug of an audit or group (depending on `type`)',
+).merge(
+  z.object({
+    type: z.enum(['audit', 'group'], {
+      description:
+        'Discriminant for reference kind, affects where `slug` is looked up',
+    }),
+    plugin: slugSchema(
+      'Plugin slug (plugin should contain referenced audit or group)',
+    ),
+  }),
+);
+export type CategoryConfigRef = z.infer<typeof categoryConfigRefSchema>;
 
 export const categoryConfigSchema = scorableSchema(
   'Category with a score calculated from audits and groups from various plugins',
-  weightedRefSchema(
-    'Weighted references to audits and/or groups for the category',
-    'Slug of an audit or group (depending on `type`)',
-  ).merge(
-    z.object({
-      type: z.enum(['audit', 'group'], {
-        description:
-          'Discriminant for reference kind, affects where `slug` is looked up',
-      }),
-      plugin: slugSchema(
-        'Plugin slug (plugin should contain referenced audit or group)',
-      ),
-    }),
-  ),
+  categoryConfigRefSchema,
   getDuplicateRefsInCategoryMetrics,
   duplicateRefsInCategoryMetricsErrorMsg,
 )
@@ -54,13 +51,15 @@ export const categoryConfigSchema = scorableSchema(
 export type CategoryConfig = z.infer<typeof categoryConfigSchema>;
 
 // helper for validator: categories have unique refs to audits or groups
-export function duplicateRefsInCategoryMetricsErrorMsg(metrics: _RefsList) {
+export function duplicateRefsInCategoryMetricsErrorMsg(
+  metrics: CategoryConfigRef[],
+) {
   const duplicateRefs = getDuplicateRefsInCategoryMetrics(metrics);
   return `In the categories, the following audit or group refs are duplicates: ${errorItems(
     duplicateRefs,
   )}`;
 }
-function getDuplicateRefsInCategoryMetrics(metrics: _RefsList) {
+function getDuplicateRefsInCategoryMetrics(metrics: CategoryConfigRef[]) {
   return hasDuplicateStrings(
     metrics.map(({ slug, type, plugin }) => `${type} :: ${plugin} / ${slug}`),
   );
