+++ packages/models/src/lib/implementation/schemas.ts
@@ -1,4 +1,5 @@
 import { z } from 'zod';
+import { MATERIAL_ICONS, MaterialIcon } from '@code-pushup/portal-client';
 import { slugRegex } from './utils';
 
 /**
@@ -150,6 +151,7 @@ export function weightedRefSchema(
     { description },
   );
 }
+export type WeightedRef = z.infer<ReturnType<typeof weightedRefSchema>>; // @TODO whi is slug optional? @matej
 
 export function scorableSchema<T extends ReturnType<typeof weightedRefSchema>>(
   description: string,
@@ -173,3 +175,8 @@ export function scorableSchema<T extends ReturnType<typeof weightedRefSchema>>(
     { description },
   );
 }
+
+export const materialIconSchema = z.enum(
+  MATERIAL_ICONS as [MaterialIcon, MaterialIcon, ...MaterialIcon[]],
+  { description: 'Icon from VSCode Material Icons extension' },
+);
