+++ packages/models/src/lib/persist-config.spec.ts
@@ -9,8 +9,8 @@ describe('persistConfigSchema', () => {
   });
 
   it('should throw if outputDir is invalid', () => {
-    const persistConfigMock = persistConfig({ outputDir: ' ' });
-
+    const persistConfigMock = persistConfig();
+    persistConfigMock.outputDir = ' ';
     expect(() => persistConfigSchema.parse(persistConfigMock)).toThrow(
       `path is invalid`,
     );
