+++ packages/models/src/lib/plugin-config-runner.ts
@@ -1,0 +1,17 @@
+import { z } from 'zod';
+import { filePathSchema } from './implementation/schemas';
+
+export const runnerConfigSchema = z.object(
+  {
+    command: z.string({
+      description: 'Shell command to execute',
+    }),
+    args: z.array(z.string({ description: 'Command arguments' })).optional(),
+    outputFile: filePathSchema('Output path'),
+  },
+  {
+    description: 'How to execute runner',
+  },
+);
+
+export type RunnerConfig = z.infer<typeof runnerConfigSchema>;
