+++ packages/models/src/lib/plugin-config.spec.ts
@@ -1,17 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { config } from '../../test';
-import { auditGroupSchema, pluginConfigSchema } from './plugin-config';
-
-describe('auditGroupSchema', () => {
-  it('should throw if a group has duplicate audit refs', () => {
-    const group = config().plugins[1].groups[0];
-    group.refs = [...group.refs, group.refs[0]];
-
-    expect(() => auditGroupSchema.parse(group)).toThrow(
-      'In plugin groups the audit refs are not unique',
-    );
-  });
-});
+import { pluginConfigSchema } from './plugin-config';
 
 describe('pluginConfigSchema', () => {
   it('should parse if plugin configuration is valid', () => {
