+++ packages/models/test/fixtures/eslint-plugin.mock.ts
@@ -1,12 +1,12 @@
 import { join } from 'node:path';
 import type {
   Audit,
-  CategoryConfig,
+  CategoryConfigRef,
   PluginConfig,
   PluginReport,
 } from '../../src';
-import { echoRunnerConfig } from './echo-runner-config.mock';
 import { ESLINT_AUDITS_MAP } from './eslint-audits.mock';
+import { echoRunnerConfig } from './runner-config.mock';
 
 const eslintMeta = {
   slug: 'eslint',
@@ -51,7 +51,7 @@ type ESLintAuditSlug = keyof typeof ESLINT_AUDITS_MAP;
 export function eslintAuditRef(
   slug: ESLintAuditSlug,
   weight = 1,
-): CategoryConfig['refs'][number] {
+): CategoryConfigRef {
   return {
     type: 'audit',
     plugin: 'eslint',
