+++ packages/models/test/fixtures/upload-config.mock.ts
@@ -1,11 +1,11 @@
-import { UploadConfig } from '../../src';
+import { UploadConfig, uploadConfigSchema } from '../../src';
 
 export function uploadConfig(opt?: Partial<UploadConfig>): UploadConfig {
-  return {
+  return uploadConfigSchema.parse({
     apiKey: 'm0ck-API-k3y',
     server: 'http://test.server.io',
     organization: 'code-pushup',
     project: 'cli',
     ...opt,
-  };
+  });
 }
